Added owner column to task list. Added delay of 5 seconds for failed logins. Raise exception when trying to destroy a Bourreau that still has tasks registered.
git-svn-id: svn+ssh://huia.bic.mni.mcgill.ca/home/cbrain/cbrain_repo/trunk@420 bcc1f7e1-216c-4ce3-888e-c5a8bb225047

+++ BrainPortal/app/controllers/bourreaux_controller.rb
@@ -141,6 +141,10 @@ class BourreauxController < ApplicationController
     
     raise "Bourreau not accessible by current user." unless @bourreau.can_be_accessed_by?(current_user)
 
+    DrmaaTask.adjust_site(@bourreau.id)
+    tasks_left = DrmaaTask.find(:all).size
+    raise "This Bourreau cannot be deleted as there are still #{tasks_left} tasks associated with it." if tasks_left > 0
+
     if @bourreau.destroy
       flash[:notice] = "Bourreau successfully deleted."
     else
